Update and fix for bug #25004
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@27257 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Foundation/NSCoder.h
@@ -222,7 +222,7 @@ extern "C" {
  *  class (decoded).  Version comes from [NSObject -getVersion].
  *  
  */
-- (unsigned int) versionForClassName: (NSString*)className;
+- (NSInteger) versionForClassName: (NSString*)className;
 
 #if OS_API_VERSION(GS_API_MACOSX, GS_API_LATEST)
 /*
