+++ Source/NSCoder.m
@@ -65,10 +65,10 @@
   return nil;
 }
 
-- (unsigned int) versionForClassName: (NSString*)className
+- (NSInteger) versionForClassName: (NSString*)className
 {
   [self subclassResponsibility: _cmd];
-  return NSNotFound;
+  return (NSInteger)NSNotFound;
 }
 
 // Encoding Data
@@ -293,8 +293,6 @@
 
 - (unsigned int) systemVersion
 {
-  //PENDING(ABR)- should probably mult major version by 1000, not 100, since,
-  //              e.g., 2.0.0 is going to be <1000
   return (((GNUSTEP_BASE_MAJOR_VERSION * 100)
     + GNUSTEP_BASE_MINOR_VERSION) * 100) + GNUSTEP_BASE_SUBMINOR_VERSION;
 }
